back/requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.1
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.7
django-appconf==1.0.4
django-bootstrap-v5==1.0.0
django-cors-headers==3.7.0
django-extensions==3.1.1
django-imagekit==4.0.2
django-seed==0.2.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
entrypoints==0.3
Faker==5.8.0
Flask==1.1.2
idna==2.10
importlib-metadata==2.1.1
inflection==0.5.1
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.7.0
itsdangerous==1.1.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
lazy-object-proxy==1.4.3
lxml==4.6.2
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
notebook==6.2.0
numpy==1.20.1
packaging==20.8
pandas==1.2.2
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
pilkit==2.0
Pillow==8.1.2
prometheus-client==0.9.0
prompt-toolkit==3.0.10
pycparser==2.20
Pygments==2.7.4
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
pywin32==300
pywinpty==0.5.7
pyzmq==21.0.1
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
terminado==0.9.2
testpath==0.4.4
text-unidecode==1.3
toml==0.10.2
tornado==6.1
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1
